Don't expose Windows SDK path as a build argument
This never worked as expected, so it's better to not expose it to users.

Previously, only some library search paths were affected by this variable.
SDK include directories etc. are always set by vcvarsall.bat, which allows us
to specify a custom SDK version, but not a custom SDK path.